requirements.txt
# Core app framework
streamlit==1.39.0

# Financial data
yfinance==0.2.43
pandas==2.2.3
numpy==1.26.4

# Visualization
matplotlib==3.9.2
plotly==5.24.1

# AI / LLM integration
openai==1.52.2

# Environment variables
python-dotenv==1.0.1

# Optional: for better HTTP performance
requests==2.32.3
mistralai>=1.0.0
